feat: refactored react structure
