+++ labellab-client/src/constants/auth.js
@@ -1,0 +1,6 @@
+export const LOGIN_REQUEST = "LOGIN_REQUEST",
+  LOGIN_SUCCESS = "LOGIN_SUCCESS",
+  LOGIN_FAILURE = "LOGIN_FAILURE",
+  SET_USER_AUTH = "SET_USER_AUTH",
+  LOGOUT_REQUEST = "LOGOUT_REQUEST",
+  LOGOUT_SUCCESS = "LOGOUT_SUCCESS";
