+++ labellab-client/src/constants/image.js
@@ -1,0 +1,5 @@
+export const POST_IMAGE_REQUEST = "POST_IMAGE_REQUEST",
+  POST_IMAGE_SUCCESS = "POST_IMAGE_SUCCESS",
+  POST_IMAGE_FAILURE = "POST_IMAGE_FAILURE",
+  IMAGE_PREVIEW_REQUEST = "IMAGE_PREVIEW_REQUEST",
+  IMAGE_PREVIEW_SUCCESS = "IMAGE_PREVIEW_SUCCESS";
