+++ labellab-client/src/constants/index.js
@@ -1,47 +1,7 @@
-export const LOGIN_REQUEST = "LOGIN_REQUEST",
-  LOGIN_SUCCESS = "LOGIN_SUCCESS",
-  LOGIN_FAILURE = "LOGIN_FAILURE",
-  REGISTER_REQUEST = "REGISTER_REQUEST",
-  REGISTER_SUCCESS = "REGISTER_SUCCESS",
-  REGISTER_FAILURE = "REGISTER_FAILURE",
-  SET_USER_AUTH = "SET_USER_AUTH",
-  LOGOUT_REQUEST = "LOGOUT_REQUEST",
-  LOGOUT_SUCCESS = "LOGOUT_SUCCESS",
-  TOKEN_TYPE = "admin",
-  SET_USER_DATA_REQUEST = "SET_USER_DATA_REQUEST",
-  SET_USER_DATA_SUCCESS = "SET_USER_DATA_SUCCESS",
-  SET_USER_DATA_FAILURE = "SET_USER_DATA_FAILURE",
-  UPLOAD_USER_IMAGE_REQUEST = "UPLOAD_USER_IMAGE_REQUEST",
-  UPLOAD_USER_IMAGE_SUCCESS = "UPLOAD_USER_IMAGE_SUCCESS",
-  UPLOAD_USER_IMAGE_FAILURE = "UPLOAD_USER_IMAGE_FAILURE",
-  INITIALIZE_PROJECT_REQUEST = "INITIALIZE_PROJECT_REQUEST",
-  INITIALIZE_PROJECT_SUCCESS = "INITIALIZE_PROJECT_SUCCESS",
-  INITIALIZE_PROJECT_FAILURE = "INITIALIZE_PROJECT_FAILURE",
-  FETCH_PROJECT_REQUEST = "FETCH_PROJECT_REQUEST",
-  FETCH_PROJECT_SUCCESS = "FETCH_PROJECT_SUCCESS",
-  FETCH_PROJECT_FAILURE = "FETCH_PROJECT_FAILURE",
-  POST_IMAGE_REQUEST = "POST_IMAGE_REQUEST",
-  POST_IMAGE_SUCCESS = "POST_IMAGE_SUCCESS",
-  POST_IMAGE_FAILURE = "POST_IMAGE_FAILURE",
-  IMAGE_PREVIEW_REQUEST = "IMAGE_PREVIEW_REQUEST",
-  IMAGE_PREVIEW_SUCCESS = "IMAGE_PREVIEW_SUCCESS",
-  FETCH_PROJECT_ALL_FAILURE = "FETCH_PROJECT_ALL_FAILURE",
-  FETCH_PROJECT_ALL_REQUEST = "FETCH_PROJECT_ALL_REQUEST",
-  FETCH_PROJECT_ALL_SUCCESS = "FETCH_PROJECT_ALL_SUCCESS",
-  POST_LABEL_REQUEST = "POST_LABEL_REQUEST",
-  POST_LABEL_SUCCESS = "POST_LABEL_SUCCESS",
-  POST_LABEL_FAILURE = "POST_LABEL_FAILURE",
-  ADD_MEMBER_REQUEST = "ADD_MEMBER_REQUEST",
-  ADD_MEMBER_SUCCESS = "ADD_MEMBER_SUCCESS",
-  ADD_MEMBER_FAILURE = "ADD_MEMBER_FAILURE",
-  UPDATE_PROJECT_REQUEST = "UPDATE_PROJECT_REQUEST",
-  UPDATE_PROJECT_SUCCESS = "UPDATE_PROJECT_SUCCESS",
-  UPDATE_PROJECT_FAILURE = "UPDATE_PROJECT_FAILURE",
-  POST_PROJECT_IMAGE_REQUEST = "POST_PROJECT_IMAGE_REQUEST",
-  POST_PROJECT_IMAGE_SUCCESS = "POST_PROJECT_IMAGE_SUCCESS",
-  POST_PROJECT_IMAGE_FAILURE = "POST_PROJECT_IMAGE_FAILURE",
-  DELETE_MEMBER_FAILURE = "DELETE_MEMBER_FAILURE",
-  DELETE_MEMBER_REQUEST = "DELETE_MEMBER_REQUEST",
-  DELETE_MEMBER_SUCCESS = "DELETE_MEMBER_SUCCESS",
-  SEARCH_PROJECTS = "SEARCH_PROJECTS",
-  SEARCH_PROJECTS_FAILURE = "SEARCH_PROJECTS_FAILURE";
+export * from "./auth";
+export * from "./register";
+export * from "./user";
+export * from "./project/index";
+export * from "./image";
+export * from "./label";
+export const TOKEN_TYPE = "admin";
