+++ labellab-client/src/constants/label.js
@@ -1,0 +1,3 @@
+export const POST_LABEL_REQUEST = "POST_LABEL_REQUEST",
+  POST_LABEL_SUCCESS = "POST_LABEL_SUCCESS",
+  POST_LABEL_FAILURE = "POST_LABEL_FAILURE";
