+++ labellab-client/src/constants/project/fetchDetails.js
@@ -1,0 +1,6 @@
+export const FETCH_PROJECT_REQUEST = "FETCH_PROJECT_REQUEST",
+  FETCH_PROJECT_SUCCESS = "FETCH_PROJECT_SUCCESS",
+  FETCH_PROJECT_FAILURE = "FETCH_PROJECT_FAILURE",
+  FETCH_PROJECT_ALL_FAILURE = "FETCH_PROJECT_ALL_FAILURE",
+  FETCH_PROJECT_ALL_REQUEST = "FETCH_PROJECT_ALL_REQUEST",
+  FETCH_PROJECT_ALL_SUCCESS = "FETCH_PROJECT_ALL_SUCCESS";
