+++ labellab-client/src/constants/project/member.js
@@ -1,0 +1,6 @@
+export const DELETE_MEMBER_FAILURE = "DELETE_MEMBER_FAILURE",
+  DELETE_MEMBER_REQUEST = "DELETE_MEMBER_REQUEST",
+  DELETE_MEMBER_SUCCESS = "DELETE_MEMBER_SUCCESS",
+  ADD_MEMBER_REQUEST = "ADD_MEMBER_REQUEST",
+  ADD_MEMBER_SUCCESS = "ADD_MEMBER_SUCCESS",
+  ADD_MEMBER_FAILURE = "ADD_MEMBER_FAILURE";
