+++ labellab-client/src/constants/project/projectDetails.js
@@ -1,0 +1,9 @@
+export const INITIALIZE_PROJECT_REQUEST = "INITIALIZE_PROJECT_REQUEST",
+  INITIALIZE_PROJECT_SUCCESS = "INITIALIZE_PROJECT_SUCCESS",
+  INITIALIZE_PROJECT_FAILURE = "INITIALIZE_PROJECT_FAILURE",
+  UPDATE_PROJECT_REQUEST = "UPDATE_PROJECT_REQUEST",
+  UPDATE_PROJECT_SUCCESS = "UPDATE_PROJECT_SUCCESS",
+  UPDATE_PROJECT_FAILURE = "UPDATE_PROJECT_FAILURE",
+  POST_PROJECT_IMAGE_REQUEST = "POST_PROJECT_IMAGE_REQUEST",
+  POST_PROJECT_IMAGE_SUCCESS = "POST_PROJECT_IMAGE_SUCCESS",
+  POST_PROJECT_IMAGE_FAILURE = "POST_PROJECT_IMAGE_FAILURE";
