+++ labellab-client/src/constants/project/search.js
@@ -1,0 +1,2 @@
+export const SEARCH_PROJECTS = "SEARCH_PROJECTS",
+  SEARCH_PROJECTS_FAILURE = "SEARCH_PROJECTS_FAILURE";
