+++ labellab-client/src/constants/register.js
@@ -1,0 +1,3 @@
+export const REGISTER_REQUEST = "REGISTER_REQUEST",
+  REGISTER_SUCCESS = "REGISTER_SUCCESS",
+  REGISTER_FAILURE = "REGISTER_FAILURE";
