+++ labellab-client/src/constants/user.js
@@ -1,0 +1,6 @@
+export const SET_USER_DATA_REQUEST = "SET_USER_DATA_REQUEST",
+  SET_USER_DATA_SUCCESS = "SET_USER_DATA_SUCCESS",
+  SET_USER_DATA_FAILURE = "SET_USER_DATA_FAILURE",
+  UPLOAD_USER_IMAGE_REQUEST = "UPLOAD_USER_IMAGE_REQUEST",
+  UPLOAD_USER_IMAGE_SUCCESS = "UPLOAD_USER_IMAGE_SUCCESS",
+  UPLOAD_USER_IMAGE_FAILURE = "UPLOAD_USER_IMAGE_FAILURE";
