+++ labellab-server/config/travis.js
@@ -1,0 +1,3 @@
+module.exports = {
+	mongoURI: "mongodb://travis:test@localhost:27017/mydb_test"
+}
